refactor: move authorizations commands, helper, and tests to cli
