+++ packages/cli/src/commands/authorizations/revoke.ts
@@ -1,7 +1,7 @@
 import color from '@heroku-cli/color'
 import {Command} from '@heroku-cli/command'
 import * as Heroku from '@heroku-cli/schema'
-import {CliUx} from '@oclif/core'
+import {Args, ux} from '@oclif/core'
 
 export default class AuthorizationsRevoke extends Command {
   static description = 'revoke OAuth authorization'
@@ -12,17 +12,19 @@ export default class AuthorizationsRevoke extends Command {
     '$ heroku authorizations:revoke 105a7bfa-34c3-476e-873a-b1ac3fdc12fb',
   ]
 
-  static args = [{name: 'id', required: true}]
+  static args = {
+    id: Args.string({required: true}),
+  }
 
   async run() {
     const {args} = await this.parse(AuthorizationsRevoke)
 
-    CliUx.ux.action.start('Revoking OAuth Authorization')
+    ux.action.start('Revoking OAuth Authorization')
 
     const {body: auth} = await this.heroku.delete<Heroku.OAuthAuthorization>(
       `/oauth/authorizations/${encodeURIComponent(args.id)}`,
     )
 
-    CliUx.ux.action.stop(`done, revoked authorization from ${color.cyan(auth.description)}`)
+    ux.action.stop(`done, revoked authorization from ${color.cyan(auth.description)}`)
   }
 }
