+++ packages/cli/test/unit/commands/authorizations/index.unit.test.ts
