+++ packages/cli/test/unit/commands/authorizations/info.unit.test.ts
