+++ packages/cli/test/unit/commands/authorizations/rotate.unit.test.ts
